Ajoute l'intégration à la base de données et la heatmap

+++ src/main/kotlin/io/jterrier/wheels/models.kt
@@ -3,6 +3,6 @@ package io.jterrier.wheels
 data class Activity(
     val id: Long,
     val name: String,
-    val distance: Double,
+    val distance: Float,
     val polyline: String
 )
